Large refactor with performance improvement
Now instead of getting the image URI from the card's JSON data, we speculatively fetch the image right away, and only rely on the JSON data if it turns out our guess was incorrect (which only happens when we are fetching a back face). This eliminates a round trip in most cases.